backend/requirements.txt
Django==5.0.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.2.2
setuptools==80.9.0
